Translate remaining static text
This does not include invisible texts (such as title attributes).
This commit also moves the LocalePicker control into i18n.js
